requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
databases[aiomysql]
sqlalchemy-utils
alembic
python-dotenv